+++ tests/doc_coverage.rs
@@ -8,6 +8,7 @@ use std::time::Duration;
 fn doc_test_coverage() {
     let mut config = Config::default();
     config.verbose = true;
+    config.force_clean = false;
     config.test_timeout = Duration::from_secs(60);
     let test_dir = get_test_path("doc_coverage");
     env::set_current_dir(&test_dir).unwrap();
@@ -34,6 +35,7 @@ fn doc_test_coverage() {
 fn doc_test_panics() {
     let mut config = Config::default();
     config.verbose = true;
+    config.force_clean = false;
     config.test_timeout = Duration::from_secs(60);
     let test_dir = get_test_path("doctest_should_panic");
     env::set_current_dir(&test_dir).unwrap();
@@ -60,6 +62,7 @@ fn doc_test_panics() {
 fn doc_test_panics_workspace() {
     let mut config = Config::default();
     config.verbose = true;
+    config.force_clean = false;
     config.test_timeout = Duration::from_secs(60);
     let test_dir = get_test_path("doctest_workspace_should_panic");
     env::set_current_dir(&test_dir).unwrap();
@@ -86,6 +89,7 @@ fn doc_test_panics_workspace() {
 fn doc_test_compile_fail() {
     let mut config = Config::default();
     config.verbose = true;
+    config.force_clean = false;
     config.test_timeout = Duration::from_secs(60);
     let test_dir = get_test_path("doctest_compile_fail_fail");
     env::set_current_dir(&test_dir).unwrap();